+++ src/aws_organization/lib/org_units.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class OrganizationalUnits(BaseModel):
+    pass
